requirements/requirements_dev.txt
Pillow==9.2.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.6.0
click==8.1.3
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==14.1.0
mypy-extensions==0.4.3
orjson==3.7.12
pathspec==0.9.0
pip==21.3.1
platformdirs==2.5.2
pyjwt==2.4.0
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
psycopg2==2.9.2
setuptools==60.2.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
wheel==0.37.1